// apps/frontend/app/api/v1/osograph/types/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & {
  [P in K]-?: NonNullable<T[P]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: string; output: string };
  JSON: { input: Record<string, unknown>; output: Record<string, unknown> };
};
// ...
export type DataConnectionAlias = {
  __typename?: "DataConnectionAlias";
  dataConnectionId: Scalars["ID"]["output"];
  datasetId: Scalars["ID"]["output"];
  id: Scalars["ID"]["output"];
  materializations: MaterializationConnection;
  modelContext?: Maybe<ModelContext>;
  orgId: Scalars["ID"]["output"];
  previewData: PreviewData;
  schema: Scalars["String"]["output"];
};
// ...
export type DataConnectionDefinition = {
  __typename?: "DataConnectionDefinition";
  /**
   * If the dataset is of type DATA_CONNECTION, this field will contain the data connection model
   * associated with the dataset.
   */
  dataConnectionAlias: DataConnectionAlias;
  datasetId: Scalars["ID"]["output"];
  orgId: Scalars["ID"]["output"];
};
// ...
export type DataIngestion = {
  __typename?: "DataIngestion";
  config: Scalars["JSON"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  datasetId: Scalars["ID"]["output"];
  factoryType: DataIngestionFactoryType;
  id: Scalars["ID"]["output"];
  materializations: MaterializationConnection;
  modelContext?: Maybe<ModelContext>;
  orgId: Scalars["ID"]["output"];
  previewData: PreviewData;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type DataIngestionDefinition = {
  __typename?: "DataIngestionDefinition";
  /**
   * The data ingestion configuration for this dataset.
   * Unlike DataModels and StaticModels, only one ingestion config is allowed per dataset.
   * Returns null if no configuration has been created yet.
   */
  dataIngestion?: Maybe<DataIngestion>;
  datasetId: Scalars["ID"]["output"];
  orgId: Scalars["ID"]["output"];
};
// ...
export type DataModelConnection = {
  __typename?: "DataModelConnection";
  edges: Array<DataModelEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

export type DataModelDefinition = {
  __typename?: "DataModelDefinition";
  /**
   * If the dataset is of type USER_MODEL, this field will contain the list of data models
   * associated with the dataset. Otherwise it will be an empty list.
   */
  dataModels: DataModelConnection;
  datasetId: Scalars["ID"]["output"];
  orgId: Scalars["ID"]["output"];
};

export type DataModelDefinitionDataModelsArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};
// ...
export type Dataset = {
  __typename?: "Dataset";
  createdAt: Scalars["DateTime"]["output"];
  creator: User;
  creatorId: Scalars["ID"]["output"];
  description?: Maybe<Scalars["String"]["output"]>;
  displayName?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  /** Whether the current user's organization is subscribed to this dataset. */
  isSubscribed: Scalars["Boolean"]["output"];
  /**
   * The materializations for this dataset. Returns all materializations regardless of their source
   * (data model, static model, or data ingestion).
   */
  materializations: MaterializationConnection;
  name: Scalars["String"]["output"];
  orgId: Scalars["ID"]["output"];
  organization: Organization;
  /**
   * The runs for this dataset. For USER_MODEL datasets, each run is related to individual data models.
   * FOR DATA_INGESTION datasets, each run is related to individual ingestion jobs.
   * For DATA_CONNECTION datasets, runs are not applicable and this field will be an empty list.
   */
  runs: RunConnection;
  tables: TableConnection;
  type: DatasetType;
  typeDefinition: DatasetTypeDefinition;
  updatedAt: Scalars["DateTime"]["output"];
};

export type DatasetIsSubscribedArgs = {
  orgId: Scalars["ID"]["input"];
};

export type DatasetMaterializationsArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};

export type DatasetRunsArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};

export type DatasetTablesArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};
// ...
export const DatasetType = {
  DataConnection: "DATA_CONNECTION",
  DataIngestion: "DATA_INGESTION",
  StaticModel: "STATIC_MODEL",
  UserModel: "USER_MODEL",
} as const;

export type DatasetType = (typeof DatasetType)[keyof typeof DatasetType];
export type DatasetTypeDefinition =
  | DataConnectionDefinition
  | DataIngestionDefinition
  | DataModelDefinition
  | StaticModelDefinition;
// ...
export type MaterializationConnection = {
  __typename?: "MaterializationConnection";
  edges: Array<MaterializationEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type Organization = {
  __typename?: "Organization";
  createdAt: Scalars["DateTime"]["output"];
  dataConnections: DataConnectionConnection;
  datasets: DatasetConnection;
  description?: Maybe<Scalars["String"]["output"]>;
  displayName?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  members: UserConnection;
  name: Scalars["String"]["output"];
  notebooks: NotebookConnection;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type RunConnection = {
  __typename?: "RunConnection";
  edges: Array<RunEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type StaticModelConnection = {
  __typename?: "StaticModelConnection";
  edges: Array<StaticModelEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};

export type StaticModelDefinition = {
  __typename?: "StaticModelDefinition";
  datasetId: Scalars["ID"]["output"];
  orgId: Scalars["ID"]["output"];
  /**
   * If the dataset is of type STATIC_MODEL, this field will contain the list of static models
   * associated with the dataset. Otherwise it will be an empty list.
   */
  staticModels: StaticModelConnection;
};

export type StaticModelDefinitionStaticModelsArgs = {
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  single?: InputMaybe<Scalars["Boolean"]["input"]>;
  where?: InputMaybe<Scalars["JSON"]["input"]>;
};
// ...
export type TableConnection = {
  __typename?: "TableConnection";
  edges: Array<TableEdge>;
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type User = {
  __typename?: "User";
  avatarUrl?: Maybe<Scalars["String"]["output"]>;
  email: Scalars["String"]["output"];
  fullName?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  organizations: OrganizationConnection;
};
// ...
export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<
  TResult,
  TParent = Record<PropertyKey, never>,
  TContext = Record<PropertyKey, never>,
  TArgs = Record<PropertyKey, never>,
> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | ResolverWithResolve<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo,
) => Promise<TResult> | TResult;
// ...
export type IsTypeOfResolverFn<
  T = Record<PropertyKey, never>,
  TContext = Record<PropertyKey, never>,
> = (
  obj: T,
  context: TContext,
  info: GraphQLResolveInfo,
) => boolean | Promise<boolean>;
// ...
export type DataConnectionDefinitionResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes["DataConnectionDefinition"] = ResolversParentTypes["DataConnectionDefinition"],
> = {
  dataConnectionAlias?: Resolver<
    ResolversTypes["DataConnectionAlias"],
    ParentType,
    ContextType
  >;
  datasetId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
// ...
export type DataIngestionDefinitionResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes["DataIngestionDefinition"] = ResolversParentTypes["DataIngestionDefinition"],
> = {
  dataIngestion?: Resolver<
    Maybe<ResolversTypes["DataIngestion"]>,
    ParentType,
    ContextType
  >;
  datasetId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
// ...
export type DataModelDefinitionResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes["DataModelDefinition"] = ResolversParentTypes["DataModelDefinition"],
> = {
  dataModels?: Resolver<
    ResolversTypes["DataModelConnection"],
    ParentType,
    ContextType,
    RequireFields<DataModelDefinitionDataModelsArgs, "first">
  >;
  datasetId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
// ...
export type DatasetResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes["Dataset"] = ResolversParentTypes["Dataset"],
> = {
  createdAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  creator?: Resolver<ResolversTypes["User"], ParentType, ContextType>;
  creatorId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  description?: Resolver<
    Maybe<ResolversTypes["String"]>,
    ParentType,
    ContextType
  >;
  displayName?: Resolver<
    Maybe<ResolversTypes["String"]>,
    ParentType,
    ContextType
  >;
  id?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  isSubscribed?: Resolver<
    ResolversTypes["Boolean"],
    ParentType,
    ContextType,
    RequireFields<DatasetIsSubscribedArgs, "orgId">
  >;
  materializations?: Resolver<
    ResolversTypes["MaterializationConnection"],
    ParentType,
    ContextType,
    RequireFields<DatasetMaterializationsArgs, "first">
  >;
  name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  organization?: Resolver<
    ResolversTypes["Organization"],
    ParentType,
    ContextType
  >;
  runs?: Resolver<
    ResolversTypes["RunConnection"],
    ParentType,
    ContextType,
    RequireFields<DatasetRunsArgs, "first">
  >;
  tables?: Resolver<
    ResolversTypes["TableConnection"],
    ParentType,
    ContextType,
    RequireFields<DatasetTablesArgs, "first">
  >;
  type?: Resolver<ResolversTypes["DatasetType"], ParentType, ContextType>;
  typeDefinition?: Resolver<
    ResolversTypes["DatasetTypeDefinition"],
    ParentType,
    ContextType
  >;
  updatedAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
};
// ...
export type StaticModelDefinitionResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes["StaticModelDefinition"] = ResolversParentTypes["StaticModelDefinition"],
> = {
  datasetId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  orgId?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
  staticModels?: Resolver<
    ResolversTypes["StaticModelConnection"],
    ParentType,
    ContextType,
    RequireFields<StaticModelDefinitionStaticModelsArgs, "first">
  >;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};